# openapi.yaml
openapi: 3.0.3
info:
  title: HackNext REST API
  description: Offline-first production hackathon platform API for collegiate engineering sprints.
  version: 1.0.0
servers:
  - url: http://localhost:5000/api
    description: Local self-hosted backend server
paths:
...
  /voting/events/{eventId}/projects:
    get:
      summary: Get public project voting gallery (strictly title and description)
      parameters:
        - name: eventId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: List of project cards with title and description only
  /voting/projects/{submissionId}/vote:
    post:
      summary: Cast strictly 1 community vote per user per event
      security:
        - BearerAuth: []
      parameters:
        - name: submissionId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Vote recorded
        '400':
          description: Already voted or self-vote attempted
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT